Fix : Change Package Structure & Rename RoleType
