+++ src/main/java/kr/co/monitoringserver/persistence/repository/UserRepository.java
@@ -1,7 +1,6 @@
 package kr.co.monitoringserver.persistence.repository;
 
 import kr.co.monitoringserver.persistence.entity.user.User;
-import kr.co.monitoringserver.service.enums.RoleType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
