+++ src/main/java/kr/co/monitoringserver/service/dtos/response/SecurityAreaLocationResDTO.java
@@ -1,7 +1,7 @@
 package kr.co.monitoringserver.service.dtos.response;
 
 import kr.co.monitoringserver.persistence.entity.Location;
-import kr.co.monitoringserver.service.enums.RoleType;
+import kr.co.monitoringserver.service.enums.UserRoleType;
 import lombok.*;
 
 import java.time.LocalTime;
@@ -22,7 +22,7 @@ public class SecurityAreaLocationResDTO {
 
         private String userName;
 
-        private RoleType userRoleType;
+        private UserRoleType userRoleType;
 
         private String userDepartment;
