+++ src/main/java/kr/co/monitoringserver/service/enums/UserRoleType.java
@@ -1,5 +1,5 @@
 package kr.co.monitoringserver.service.enums;
 
-public enum RoleType {
+public enum UserRoleType {
     USER1, USER2, USER3, ADMIN
 }
