+++ src/main/java/kr/co/monitoringserver/service/service/fcm/FirebaseCloudMessageService.java
@@ -1,10 +1,10 @@
-package kr.co.monitoringserver.service.service;
+package kr.co.monitoringserver.service.service.fcm;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.common.net.HttpHeaders;
-import kr.co.monitoringserver.service.dtos.request.FCMMessageDTO;
+import kr.co.monitoringserver.service.dtos.request.fcm.FCMMessageDTO;
 import lombok.RequiredArgsConstructor;
 import okhttp3.*;
 import org.springframework.core.io.ClassPathResource;
